refactor into two workspaces
