+++ dist-tracing/src/telemetry.rs
@@ -1,56 +1,10 @@
 use crate::trace::{Event, Span};
-use crate::visitor::{event_to_values, span_to_values, HoneycombVisitor};
-use libhoney::FieldHolder;
-use std::collections::HashMap;
-use std::sync::Mutex;
 
 pub trait Telemetry {
     type Visitor: Default + tracing::field::Visit;
 
     fn report_span<'a>(&self, span: Span<'a, Self::Visitor>);
     fn report_event<'a>(&self, event: Event<'a, Self::Visitor>);
-}
-
-pub struct HoneycombTelemetry {
-    honeycomb_client: Mutex<libhoney::Client<libhoney::transmission::Transmission>>,
-}
-
-impl HoneycombTelemetry {
-    pub fn new(cfg: libhoney::Config) -> Self {
-        let honeycomb_client = libhoney::init(cfg);
-
-        // publishing requires &mut so just mutex-wrap it
-        // FIXME: may not be performant, investigate options (eg mpsc)
-        let honeycomb_client = Mutex::new(honeycomb_client);
-
-        HoneycombTelemetry { honeycomb_client }
-    }
-
-    fn report_data(&self, data: HashMap<String, ::libhoney::Value>) {
-        // succeed or die. failure is unrecoverable (mutex poisoned)
-        let mut client = self.honeycomb_client.lock().unwrap();
-        let mut ev = client.new_event();
-        ev.add(data);
-        let res = ev.send(&mut client);
-        if let Err(err) = res {
-            // unable to report telemetry (buffer full) so log msg to stderr
-            // TODO: figure out strategy for handling this (eg report data loss event)
-            eprintln!("error sending event to honeycomb, {:?}", err);
-        }
-    }
-}
-
-impl Telemetry for HoneycombTelemetry {
-    type Visitor = HoneycombVisitor;
-    fn report_span(&self, span: Span<Self::Visitor>) {
-        let data = span_to_values(span);
-        self.report_data(data);
-    }
-
-    fn report_event(&self, event: Event<Self::Visitor>) {
-        let data = event_to_values(event);
-        self.report_data(data);
-    }
 }
 
 #[derive(Default)]
@@ -72,6 +26,7 @@ impl Telemetry for BlackholeTelemetry {
 #[cfg(test)]
 pub(crate) mod test {
     use super::*;
+    use std::sync::Mutex;
     use std::sync::Arc;
 
     /// Mock telemetry capability
