+++ honeycomb-tracing/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "honeycomb-tracing"
+version = "0.1.0"
+authors = ["Inanna Malick <inanna@recursion.wtf>"]
+edition = "2018"
+
+# FIXME:
+description = "provides a tracing subscriber that publishes spans and events to honeycomb.io"
+documentation = "https://pkinsky.github.io/honeycomb-tracing/honeycomb_tracing/"
+repository = "https://github.com/inanna-malick/honeycomb-tracing"
+keywords = ["tracing", "honeycomb", "instrumentation"]
+license = "MIT"
+readme = "README.md"
+
+[dependencies]
+tracing = "0.1.12"
+tracing-core = "0.1.9"
+dist-tracing =  { path = "../dist-tracing" }
+libhoney-rust = "0.1.3"
+
+[dev-dependencies]
+tracing-attributes = "0.1.5"
+futures-preview = { version = "0.3.0-alpha.19", features = ["compat"] }
+tracing-subscriber = "0.2.0"
+tokio = { version = "0.2", features = ["full"] }
+tracing-futures = "0.2.1"
