+++ honeycomb-tracing/src/lib.rs
@@ -1,0 +1,7 @@
+#![deny(warnings)]
+mod telemetry;
+mod visitor;
+
+pub use crate::telemetry::HoneycombTelemetry;
+pub use crate::visitor::HoneycombVisitor;
+pub use dist_tracing::{TelemetryLayer, TraceCtx};
