+++ honeycomb-tracing/src/telemetry.rs
@@ -1,0 +1,47 @@
+use dist_tracing::{Event, Span, Telemetry};
+use crate::visitor::{event_to_values, span_to_values, HoneycombVisitor};
+use libhoney::FieldHolder;
+use std::collections::HashMap;
+use std::sync::Mutex;
+
+pub struct HoneycombTelemetry {
+    honeycomb_client: Mutex<libhoney::Client<libhoney::transmission::Transmission>>,
+}
+
+impl HoneycombTelemetry {
+    pub fn new(cfg: libhoney::Config) -> Self {
+        let honeycomb_client = libhoney::init(cfg);
+
+        // publishing requires &mut so just mutex-wrap it
+        // FIXME: may not be performant, investigate options (eg mpsc)
+        let honeycomb_client = Mutex::new(honeycomb_client);
+
+        HoneycombTelemetry { honeycomb_client }
+    }
+
+    fn report_data(&self, data: HashMap<String, ::libhoney::Value>) {
+        // succeed or die. failure is unrecoverable (mutex poisoned)
+        let mut client = self.honeycomb_client.lock().unwrap();
+        let mut ev = client.new_event();
+        ev.add(data);
+        let res = ev.send(&mut client);
+        if let Err(err) = res {
+            // unable to report telemetry (buffer full) so log msg to stderr
+            // TODO: figure out strategy for handling this (eg report data loss event)
+            eprintln!("error sending event to honeycomb, {:?}", err);
+        }
+    }
+}
+
+impl Telemetry for HoneycombTelemetry {
+    type Visitor = HoneycombVisitor;
+    fn report_span(&self, span: Span<Self::Visitor>) {
+        let data = span_to_values(span);
+        self.report_data(data);
+    }
+
+    fn report_event(&self, event: Event<Self::Visitor>) {
+        let data = event_to_values(event);
+        self.report_data(data);
+    }
+}
